extracted class init for method lifecycle from class-decor
