+++ src/mixins/disposable-decorator.ts
@@ -1,5 +1,5 @@
 import { Disposers } from "../utils/disposers";
-import {after, onInstance, chain} from "../utils/class-decor";
+import {after, onInstance, chain} from "../utils/class-decor/";
 export interface DisposableCompMixin extends React.Component{
     readonly disposer: Disposers;
 }
